+++ static/js/settings.js
@@ -98,11 +98,11 @@ exports.build_page = function () {
         settings_label: exports.settings_label,
         demote_inactive_streams_values: settings_config.demote_inactive_streams_values,
         twenty_four_hour_time_values: settings_config.twenty_four_hour_time_values,
-        general_settings: settings_notifications.all_notifications.general_settings,
-        notification_settings: settings_notifications.all_notifications.settings,
+        general_settings: settings_config.all_notifications().general_settings,
+        notification_settings: settings_config.all_notifications().settings,
         desktop_icon_count_display_values: settings_notifications.desktop_icon_count_display_values,
         show_push_notifications_tooltip:
-            settings_notifications.all_notifications.show_push_notifications_tooltip,
+            settings_config.all_notifications().show_push_notifications_tooltip,
         display_settings: settings_config.get_all_display_settings(),
         user_can_change_name: settings_account.user_can_change_name(),
         user_can_change_avatar: settings_account.user_can_change_avatar(),
